refactor(auth): move verify to openid/delegation module
