+++ src/libs/auth/src/openid/delegation/mod.rs
@@ -1,2 +1,3 @@
 mod impls;
 pub mod types;
+pub mod verify;
