+++ src/libs/auth/src/openid/mod.rs
@@ -4,6 +4,5 @@ pub mod jwkset;
 pub mod jwt;
 pub mod types;
 mod utils;
-mod verify;
 
-pub use verify::*;
+pub use delegation::verify::*;
